@@ -15,7 +15,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+2
+version: 1.0.0+6
 
 environment:
   sdk: ">=2.12.0 <3.0.0"
@@ -30,29 +30,31 @@ dependencies:
   hive_flutter: ^1.0.0
   google_fonts: ^2.1.0
   at_common_flutter: ^2.0.2
-  at_onboarding_flutter: ^2.1.2
-  at_client_mobile: ^2.0.3
-  at_client: ^2.0.2
-  at_contact: ^2.0.0
-  at_lookup: ^2.0.1
-  at_contacts_flutter: ^2.0.3
+  at_onboarding_flutter: ^3.0.0
+  at_client_mobile: ^3.0.1
+  at_client: ^3.0.0
+  at_contact: ^3.0.0
+  at_lookup: ^3.0.0
+  at_contacts_flutter: ^3.0.0
   flutter_image_compress: ^1.0.0
   path_provider: ^2.0.1
   permission_handler: ^8.1.3
-  at_commons: ^2.0.1
+  at_commons: ^3.0.0
   at_base2e15: ^1.0.0
   youtube_player_flutter: ^8.0.0
   receive_sharing_intent: ^1.4.5
   http: ^0.13.3
-  file_picker: ^3.0.2+2
+  file_picker: ^4.0.2
   geocoder: ^0.2.1
   easy_debounce: ^2.0.0
   validators: ^3.0.0
-  at_follows_flutter: ^2.0.4
+  at_follows_flutter: ^3.0.0
   latlong2: ^0.8.0
-  at_location_flutter: ^2.0.4
   showcaseview: ^1.1.1
   video_compress: ^3.1.0
+  at_location_flutter: ^3.0.0
+
+  html_editor_enhanced: ^2.3.0
   flutter_widget_from_html: ^0.6.2
   shared_preferences: ^2.0.7
   visibility_detector: ^0.2.0
@@ -65,12 +67,18 @@ dependencies:
   desktop_window: ^0.4.0
   url_launcher: ^6.0.12
 
+dependency_overrides:
+  at_onboarding_flutter:
+   git:
+     url: https://github.com/atsign-foundation/at_widgets.git
+     path: at_onboarding_flutter
+     ref: trunk
+  flutter_colorpicker: ^0.6.0 #because html_editor_enhanced uses 0.4.0, which gives flutter event errors
+  file_picker: ^4.0.0
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
-dependency_overrides:
-  file_picker: ^4.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
