@@ -95,13 +95,13 @@ extension FieldsEnumValues on FieldsEnum {
         return 'Last Name';
 
       case FieldsEnum.PHONE:
-        return 'Phone';
+        return 'Phone Number';
 
       case FieldsEnum.EMAIL:
-        return 'Email';
+        return 'Email Address';
 
       case FieldsEnum.ABOUT:
-        return 'Write a short bio or quote';
+        return 'About';
 
       case FieldsEnum.PRONOUN:
         return 'Preferred Pronoun';
@@ -113,13 +113,19 @@ extension FieldsEnumValues on FieldsEnum {
         return 'Type postal code, city, country, or street';
 
       case FieldsEnum.TWITTER:
+        return 'Twitter';
       case FieldsEnum.INSTAGRAM:
+        return 'Instagram';
       case FieldsEnum.FACEBOOK:
+        return 'Facebook';
       case FieldsEnum.LINKEDIN:
+        return 'Linkedin';
       case FieldsEnum.TUMBLR:
+        return 'Tumblr';
       case FieldsEnum.MEDIUM:
+        return 'Medium';
       case FieldsEnum.YOUTUBE:
-        return 'USERNAME';
+        return 'Youtube';
 
       case FieldsEnum.PS4:
         return 'PS4';
@@ -328,7 +334,7 @@ extension value on RootEnvironment {
 }
 
 enum CustomContentStatus { Exists, Success, Fails }
-enum CustomContentType { Text, Link, Number, Image, Youtube }
+enum CustomContentType { Text, Link, Number, Image, Youtube, Html, Location }
 
 extension values on CustomContentType {
   String get name {
@@ -347,6 +353,12 @@ extension values on CustomContentType {
 
       case CustomContentType.Youtube:
         return 'Youtube';
+
+      case CustomContentType.Html:
+        return 'Html';
+
+      case CustomContentType.Location:
+        return 'Location';
 
       default:
         return '';
@@ -396,6 +408,9 @@ CustomContentType customContentNameToType(String name) {
 
     case 'Youtube':
       return CustomContentType.Youtube;
+
+    case 'Html':
+      return CustomContentType.Html;
 
     default:
       return CustomContentType.Text;
